// frontend/src/screens/userProfile/redux/types.ts
export interface userProfileState {
  email: string;
  firstName: string;
  lastName: string;
  address: string;
  city: string;
  pinCode: string;
  phoneNumber: string;
  userId: string;
  profileUrl: string;
  isEdit: boolean;
  onesignalUserId: string;
  latitude: string;
  longitude: string;
}